Factor out jintrac data handling
